% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invites.R
\name{ol_invite_send_link}
\alias{ol_invite_send_link}
\title{Send Invite Links}
\usage{
ol_invite_send_link(con, email)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{email}{email of a user}
}
\description{
Send Invite Links
}
\examples{
if (interactive()) ol_invite_send_link(onelogin(), "example@example.com")
}
