% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{ol_role_get_by_id}
\alias{ol_role_get_by_id}
\title{Get details of particular role}
\usage{
ol_role_get_by_id(con, role_id)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{role_id}{the id of a role}
}
\value{
a tibble of details on the role
}
\description{
Get details of particular role
}
\examples{
if (interactive()) ol_role_get_by_id(onelogin(), 1234)
}
