% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{ol_roles_get}
\alias{ol_roles_get}
\title{Get all available roles}
\usage{
ol_roles_get(con, ...)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{...}{filters for the roles. See options in the \href{https://developers.onelogin.com/api-docs/1/roles/get-roles}{API docs}.}
}
\value{
a tibble of roles
}
\description{
Get all available roles
}
\examples{
if (interactive()) ol_roles_get(onelogin())
}
