% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{ol_token_get}
\alias{ol_token_get}
\title{Generate a 'OneLogin' token}
\usage{
ol_token_get(con)
}
\arguments{
\item{con}{a 'OneLogin' connection}
}
\value{
A 'OneLogin' connection with auth token
}
\description{
Generate a 'OneLogin' token
}
\examples{
if(interactive()) ol_token_get(onelogin())
}
