% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{ol_token_get_rate_limit}
\alias{ol_token_get_rate_limit}
\title{Get Token Rate Limit}
\usage{
ol_token_get_rate_limit(con)
}
\arguments{
\item{con}{a onelogin object}
}
\value{
a onelogin connection
}
\description{
Get Token Rate Limit
}
