% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{ol_token_refresh}
\alias{ol_token_refresh}
\title{Refresh a OneLogin token}
\usage{
ol_token_refresh(con)
}
\arguments{
\item{con}{a onelogin object}
}
\value{
a onelogin connection
}
\description{
Refresh a OneLogin token
}
