% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{ol_token_revoke}
\alias{ol_token_revoke}
\title{Revoke \code{OneLogin} access token}
\usage{
ol_token_revoke(con)
}
\arguments{
\item{con}{a 'OneLogin' connection}
}
\value{
A tibble of response status
}
\description{
Revoke \code{OneLogin} access token
}
\examples{
#' if(interactive()) ol_token_revoke(onelogin())
}
