% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_assign_role}
\alias{ol_user_assign_role}
\title{Assign role to user}
\usage{
ol_user_assign_role(con, user_id, role_id_array)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{role_id_array}{numeric, one or more roles}
}
\description{
Assign role to user
}
\examples{
if (interactive()) ol_user_assign_role(onelogin(), 54963040, 268986)
}
