% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_create}
\alias{ol_user_create}
\title{Create a 'OneLogin' user.}
\usage{
ol_user_create(con, firstname, lastname, email, username, ...)
}
\arguments{
\item{con}{a onelogin connection}

\item{firstname}{first name, character}

\item{lastname}{last name, character}

\item{email}{full email, character}

\item{username}{username}

\item{...}{other named parameters for the person}
}
\value{
tibble of data returned by API call
}
\description{
For a full listing of available fields, see the \link{API documentation}\url{https://developers.onelogin.com/api-docs/1/users/create-user}
}
