% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_apps}
\alias{ol_user_get_apps}
\title{Get Apps for User}
\usage{
ol_user_get_apps(con, user_id)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}
}
\value{
A tibble of user data (one row)
}
\description{
Get Apps for User
}
\examples{
if (interactive()) ol_user_get_apps(onelogin(), 54400533)
}
