% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_by_id}
\alias{ol_user_get_by_id}
\title{Get a User by their ID}
\usage{
ol_user_get_by_id(con, user_id)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}
}
\value{
A tibble of user data (one row)
}
\description{
Get a User by their ID
}
\examples{
if (interactive()) ol_user_get_by_id(onelogin(), 54400533)
}
