% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_custom_fields}
\alias{ol_user_get_custom_fields}
\title{Get Custom fields available for users}
\usage{
ol_user_get_custom_fields(con)
}
\value{
tibble of custom fields available
}
\description{
Get Custom fields available for users
}
