% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_roles}
\alias{ol_user_get_roles}
\title{Get Roles for a User}
\usage{
ol_user_get_roles(con, user_id)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}
}
\value{
A data frame of the user id and role
}
\description{
Get Roles for a User
}
\examples{
if (interactive()) ol_user_get_roles(onelogin(), 54400533)
}
