% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_lock_account}
\alias{ol_user_lock_account}
\title{Lock user}
\usage{
ol_user_lock_account(con, user_id, locked_until = 0)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{locked_until}{numeric, number of minutes, default to 0 see \href{https://developers.onelogin.com/api-docs/1/users/lock-user-account}{API documentation} for more details}
}
\description{
Lock user
}
\examples{
if (interactive()) ol_user_lock_account(onelogin(), 54963040)
}
