% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_log_out}
\alias{ol_user_log_out}
\title{Log user out}
\usage{
ol_user_log_out(con, user_id)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}
}
\description{
Log user out
}
\examples{
if (interactive()) ol_user_log_out(onelogin(), 54963040)
}
