% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_pwd_cleartext}
\alias{ol_user_pwd_cleartext}
\title{Title}
\usage{
ol_user_pwd_cleartext(con, id, password, password_confirmation,
  validate_policy = FALSE)
}
\arguments{
\item{validate_policy}{}
}
\description{
Title
}
