% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_set_custom_attr}
\alias{ol_user_set_custom_attr}
\title{Set custom attribute for users}
\usage{
ol_user_set_custom_attr(con, user_id, ...)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{...}{named custom attributes to set}
}
\description{
See \code{\link{ol_user_get_custom_fields}} to get custom fields
}
\examples{
if (interactive()) ol_user_set_custom_attr(onelogin, 54963040, attr = "value")
}
