% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_set_state}
\alias{ol_user_set_state}
\title{Set user state}
\usage{
ol_user_set_state(con, user_id, state)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{state}{numeric, 0-3, see \href{https://developers.onelogin.com/api-docs/1/users/set-state}{documentation} for value meanings}
}
\description{
Set user state
}
\examples{
if (interactive()) ol_user_set_state(onelogin(), 54963040, 1)
}
