% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_update}
\alias{ol_user_update}
\title{Update user information by ID}
\usage{
ol_user_update(con, user_id, ...)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{...}{named parameters to change in request}
}
\value{
A tibble of user data
}
\description{
Update user information by ID
}
\examples{
if (interactive()) ol_user_update(onelogin(), 54963040, firstname = "Fake1")
}
