% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_users_get}
\alias{ol_users_get}
\title{Get Users from OneLogin}
\usage{
ol_users_get(con, ...)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{...}{filter parameters, optional; see API documentation}
}
\value{
A tibble of users and their attributes
}
\description{
You can filter the user by various parameters in onelogin. See the
\href{https://developers.onelogin.com/api-docs/1/users/get-users}{page in the API docs} for filter options.
}
\details{
The id column in the returned tibble is the user_id for any of the user
functions that are by id.
}
\examples{
if (interactive()) ol_users_get(onelogin())
if (interactive()) ol_users_get(onelogin(), firstname = "name")
}
