% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invites.R
\name{ol_invite_gen_link}
\alias{ol_invite_gen_link}
\title{Generate Invite Link}
\usage{
ol_invite_gen_link(con, email)
}
\arguments{
\item{email}{email of a user}
}
\value{
data frame with email and invite link
}
\description{
Generate Invite Link
}
