% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{ol_token_revoke}
\alias{ol_token_revoke}
\title{Revoke a OneLogin Connection Token}
\usage{
ol_token_revoke(con)
}
\arguments{
\item{con}{a onelogin object}
}
\value{
a onelogin connection
}
\description{
Revoke a OneLogin Connection Token
}
