% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_by_id}
\alias{ol_user_get_by_id}
\title{Get a User by their ID}
\usage{
ol_user_get_by_id(con, id)
}
\arguments{
\item{con}{a onelogin object}

\item{id}{user id}
}
\value{
a tibble of user data
}
\description{
Get a User by their ID
}
