% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_custom_fields}
\alias{ol_user_get_custom_fields}
\title{Get custom fields available for users}
\usage{
ol_user_get_custom_fields(con)
}
\arguments{
\item{con}{a 'OneLogin' connection}
}
\value{
A tibble of custom fields available
}
\description{
Get custom fields available for users
}
\examples{
if (interactive()) ol_user_get_custom_fields(onelogin())
}
