% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_get_roles}
\alias{ol_user_get_roles}
\title{Get Roles for a User}
\usage{
ol_user_get_roles(con, id)
}
\arguments{
\item{id}{user id}
}
\value{
a data frame of the user id and role
}
\description{
Get Roles for a User
}
