% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_pwd_cleartext}
\alias{ol_user_pwd_cleartext}
\title{Set or change user's password}
\usage{
ol_user_pwd_cleartext(con, user_id, password, password_confirmation,
  validate_policy = FALSE)
}
\arguments{
\item{con}{a 'OneLogin' connection}

\item{user_id}{user id}

\item{password}{character, new password}

\item{password_confirmation}{character, new password}

\item{validate_policy}{utilize password policy checks? defaults to FALSE}
}
\description{
Set or change user's password
}
\examples{
if (interactive()) ol_user_pwd_cleartext(onelogin(), 54963040, "pwd", "pwd")
}
