% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_user_update}
\alias{ol_user_update}
\title{Update User Information by ID}
\usage{
ol_user_update(con, id, ...)
}
\arguments{
\item{con}{a onelogin object}

\item{id}{user id}

\item{...}{named parameters to change in request}
}
\value{
tibble of user data
}
\description{
Update User Information by ID
}
