% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ol_users_get}
\alias{ol_users_get}
\title{Get Users from OneLogin}
\usage{
ol_users_get(con, ...)
}
\arguments{
\item{...}{filter parameters, optional; see API documentation}
}
\value{
a tibble of users and their attributes
}
\description{
You can filter the user by various parameters in onelogin. See the
\href{https://developers.onelogin.com/api-docs/1/users/get-users}{page in the API docs} for filter options.
}
