% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onelogin.R
\docType{data}
\name{onelogin}
\alias{onelogin}
\alias{ONELOGIN}
\title{Class representing a Onelogin API client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ONELOGIN
}
\description{
Class representing a Onelogin API client
}
\section{Usage}{

\preformatted{
client <- onelogin$new(host = 'https://api.us.onelogin.com/',
  client_id = 'my_id',
  client_secret = 'my_secret')
}
}

\section{Details}{


This class allows a user to manipulate \href{onelogin.com}{onelogin} via the
\href{https://developers.onelogin.com/}{onelogin API}.

Authentication is done by providing a client ID and secret to get both an
auth and refresh token.
}

\keyword{datasets}
